requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
sqlalchemy==2.0.37
alembic==1.14.1
mysqlclient==2.2.7
python-multipart==0.0.20
pillow==11.1.0
python-jose==3.3.0
boto3==1.36.6
uuid==1.30